OSE-523 - Switched architecture to use notify
