+++ terraform-modules/lambda-stats/code/stats/stats.py
@@ -1,58 +1,14 @@
-from __future__ import print_function
-import json
-import os
-import requests
-from utils.utils_dates import last_month_start
-
 from utils.utils_db import count_previous_month, count_previous_year, db_get_table_name
 from utils.utils_db_ips import db_count_items
-
-
-def build_markdown_block(text):
-    return {"type": "section", "text": {"type": "mrkdwn", "text": text}}
-
-
-def build_message():
-    last_month_count = count_previous_month()
-    last_year = count_previous_year()
-    total = db_count_items(db_get_table_name())
-
-    last_month = last_month_start()
-    last_month_year_text = last_month.strftime("%B %Y")
-    last_year_text = last_month.strftime("%Y")
-
-    blocks = [
-        build_markdown_block(f"Total new findings for {last_month_year_text}: *{last_month_count}*"),
-        build_markdown_block(f"Total new findings for {last_year_text}: *{last_year}*"),
-        build_markdown_block(f"Total findings all time: *{total}*"),
-    ]
-    return {"blocks": blocks}
+from utils.utils_aws import publish_to_sns
 
 
 def lambda_handler(event, context):  # pylint:disable=unused-argument
 
-    slack_url = os.environ["SLACK_WEBHOOK_URL"]
-    slack_channel = os.environ["SLACK_CHANNEL"]
-    slack_username = os.environ["SLACK_USERNAME"]
-    slack_emoji = os.environ["SLACK_EMOJI"]
-    subject = "Domain Protect Monthly Stats"
-
-    payload = {
-        "channel": slack_channel,
-        "username": slack_username,
-        "icon_emoji": slack_emoji,
-        "attachments": [build_message()],
-        "text": subject,
+    json_data = {
+        "LastMonth": count_previous_month(),
+        "LastYear": count_previous_year(),
+        "Total": db_count_items(db_get_table_name()),
     }
 
-    response = requests.post(
-        slack_url,
-        data=json.dumps(payload),
-        headers={"Content-Type": "application/json"},
-    )
-
-    if response.status_code != 200:
-        ValueError(f"Request to Slack returned error {response.status_code}:\n{response.text}")
-
-    else:
-        print(f"Stats message sent to {slack_channel} Slack channel")
+    publish_to_sns(json_data, "Monthly Stats")
