+++ terraform-modules/lambda-stats/main.tf
@@ -27,10 +27,9 @@ resource "aws_lambda_function" "lambda" {
   # checkov:skip=CKV_AWS_115: concurrency limit on individual Lambda function not required
   # checkov:skip=CKV_AWS_117: not configured inside VPC as no handling of confidential data
 
-  count            = length(var.slack_channels)
   filename         = "${path.module}/build/stats.zip"
-  function_name    = "${var.project}-stats-slack-${element(var.slack_channels, count.index)}-${local.env}"
-  description      = "${var.project} Lambda function posting stats to ${element(var.slack_channels, count.index)} Slack channel"
+  function_name    = "${var.project}-stats-${local.env}"
+  description      = "${var.project} Lambda function posting stats to SNS"
   role             = var.lambda_role_arn
   handler          = "stats.lambda_handler"
   kms_key_arn      = var.kms_arn
@@ -42,12 +41,12 @@ resource "aws_lambda_function" "lambda" {
 
   environment {
     variables = {
-      SLACK_CHANNEL       = element(var.slack_channels, count.index)
-      SLACK_WEBHOOK_URL   = element(var.slack_webhook_urls, count.index)
-      SLACK_EMOJI         = var.slack_emoji
-      SLACK_USERNAME      = var.slack_username
-      PROJECT             = var.project
-      TERRAFORM_WORKSPACE = local.env
+      PROJECT                  = var.project
+      TERRAFORM_WORKSPACE      = local.env
+      ORG_PRIMARY_ACCOUNT      = var.org_primary_account
+      SECURITY_AUDIT_ROLE_NAME = var.security_audit_role_name
+      EXTERNAL_ID              = var.external_id
+      SNS_TOPIC_ARN            = var.sns_topic_arn
     }
   }
 
@@ -61,29 +60,26 @@ resource "aws_lambda_function" "lambda" {
 }
 
 resource "aws_lambda_alias" "lambda" {
-  count            = length(var.slack_channels)
-  name             = "${var.project}-slack-${var.slack_channels[count.index]}-${local.env}"
-  description      = "Alias for ${var.project}-${element(var.slack_channels, count.index)}-${local.env}"
-  function_name    = aws_lambda_function.lambda[count.index].function_name
+  name             = "${var.project}-stats-${local.env}"
+  description      = "Alias for ${var.project}-stats-${local.env}"
+  function_name    = aws_lambda_function.lambda.function_name
   function_version = "$LATEST"
 }
 
 resource "aws_cloudwatch_event_rule" "first_day_of_month" {
   name                = "${var.project}-stats-${local.env}"
   description         = "Triggers ${var.project} lambda stats function according to schedule"
-  schedule_expression = "cron(0 9 1 * ? *)" # 9am on the first of the month
+  schedule_expression = var.schedule_experession
 }
 
 resource "aws_cloudwatch_event_target" "run_lambda_on_first" {
-  count = length(var.slack_channels)
-  rule  = aws_cloudwatch_event_rule.first_day_of_month.name
-  arn   = aws_lambda_function.lambda[count.index].arn
+  rule = aws_cloudwatch_event_rule.first_day_of_month.name
+  arn  = aws_lambda_function.lambda.arn
 }
 
 resource "aws_lambda_permission" "allow_cloudwatch_to_call_stats" {
-  count         = length(var.slack_channels)
   action        = "lambda:InvokeFunction"
-  function_name = aws_lambda_function.lambda[count.index].function_name
+  function_name = aws_lambda_function.lambda.function_name
   principal     = "events.amazonaws.com"
   source_arn    = aws_cloudwatch_event_rule.first_day_of_month.arn
 }
